Fixed mistake in fault localisation
In the fault localisation algorithm there was a mistake making it
too optimistic. The function find_fault_inditcators returns edges
that has fault indicators. Preiously the code checked if there
were fault indicators on both sides of the faulted edge using the
edge vertices of the edge containing the fault instead of the
actual location of the fault indicators.